Driving Gear 1
The car now successfuly drives around the tracks, remaining in gear one. I inspired myself off of the TORCS driving tutorials, doing some calculus to make the car steer clear of the sides of the track while maintaining itself on the middle. It still requires work, especially to brake and shift gears.